+++ bt_ws/__pycache__/compare_node.cpython-310.pyc
