+++ bt_ws/compare_node.py
