+++ bt_ws/test_compare.py
