--- demo_ws/index.html
@@ -1,211 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8">
-  <title>Waypoint & Comparison Grid</title>
-  <style>
-    @import url('https://fonts.googleapis.com/css2?family=Comic+Neue:wght@400;700&display=swap');
-    body {
-      font-family: 'Comic Neue', cursive;
-      background: linear-gradient(135deg, #FFDEE9 0%, #B5FFFC 100%);
-      margin: 0;
-      padding: 20px;
-      color: #333;
-    }
-    h1 {
-      text-align: center;
-      font-size: 2.2em;
-      margin-bottom: 20px;
-      color: #FF6F91;
-      text-shadow: 1px 1px 3px rgba(0,0,0,0.1);
-    }
-    #grid {
-      display: grid;
-      /* 3열로 배치. 필요에 따라 repeat(n, 1fr) 값 변경 가능 */
-      grid-template-columns: repeat(3, 1fr);
-      gap: 20px 10px;
-    }
-    .cell {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      background: #FFF;
-      border-radius: 12px;
-      box-shadow: 0 4px 6px rgba(0,0,0,0.1);
-      padding: 10px;
-      transition: transform 0.2s;
-    }
-    .cell:hover {
-      transform: translateY(-5px);
-    }
-    .cell h3 {
-      margin: 0 0 8px;
-      font-size: 1.1em;
-      opacity: 0.5;
-      transition: opacity .2s;
-      color: #555;
-      text-align: center;
-    }
-    .cell img {
-      width: 100%;
-      max-height: 180px;
-      object-fit: contain;
-      border: 2px dashed #FF9FF3;
-      border-radius: 8px;
-      display: none;
-      background: #FFF;
-    }
-    .cell .noimage {
-      font-size: 0.9em;
-      color: #888;
-      display: none;
-    }
-    /* 이미지 로드 완료 상태 */
-    .cell.ready h3 {
-      opacity: 1;
-      color: #FF6F91;
-    }
-    .cell.ready img {
-      display: block;
-    }
-    .cell.ready .noimage {
-      display: none;
-    }
-    /* 이미지 없음 상태 */
-    .cell.not-ready h3 {
-      opacity: 0.5;
-    }
-    .cell.not-ready img {
-      display: none;
-    }
-    .cell.not-ready .noimage {
-      display: block;
-    }
-  </style>
-</head>
-<body>
-  <h1>Waypoint & Comparison Grid</h1>
-  <div id="grid"></div>
-
-  <script>
-    // Waypoint 이미지 개수: waypoint_1.png ~ waypoint_NUM_ENTRIES.png 가 준비되어 있어야 함.
-    const NUM_ENTRIES = 5; // 예: 5개. 실제 Waypoint 이미지 개수에 맞춰 변경하세요.
-
-    // 표시할 타입 목록
-    const TYPES = ['Reference', 'Current', 'Diff'];
-
-    // 타입별 파일명 prefix 매핑
-    const prefixMap = {
-      'Reference': 'reference',
-      'Current':   'waypoint',
-      'Diff':      'comparison_result'
-    };
-
-    // suffixIndex 계산 함수
-    function getSuffixIndex(type, idx) {
-      if (type === 'Reference') {
-        // Reference 이미지는 0부터: reference_0.png, reference_1.png, ...
-        return idx - 1;
-      }
-      else if (type === 'Current') {
-        // Waypoint(Current)은 1부터: waypoint_1.png, waypoint_2.png, ...
-        return idx;
-      }
-      else if (type === 'Diff') {
-        // Comparison Result(Diff)는 0부터: comparison_result_0.png, comparison_result_1.png, ...
-        return idx - 1;
-      }
-      // 안전 fallback
-      return idx;
-    }
-
-    const grid = document.getElementById('grid');
-
-    // 1) 셀 생성: Waypoint 인덱스 i를 1..NUM_ENTRIES까지 순회
-    for (let i = 1; i <= NUM_ENTRIES; i++) {
-      TYPES.forEach(type => {
-        const cell = document.createElement('div');
-        cell.classList.add('cell', 'not-ready');
-
-        // suffixIndex 및 제목 텍스트 결정
-        const suffixIndex = getSuffixIndex(type, i);
-        const title = document.createElement('h3');
-        let titleText;
-        if (type === 'Reference') {
-          titleText = `Reference ${suffixIndex}`;
-        } else if (type === 'Current') {
-          titleText = `Waypoint ${suffixIndex}`;
-        } else { // Diff
-          titleText = `Comparison Result ${suffixIndex}`;
-        }
-        title.textContent = titleText;
-        cell.appendChild(title);
-
-        // img 요소: id를 prefix_suffixIndex 형태로 지정
-        const img = document.createElement('img');
-        const prefix = prefixMap[type];
-        const imgId = `${prefix}_${suffixIndex}`;
-        img.id = imgId;
-        img.alt = titleText;
-        cell.appendChild(img);
-
-        // “No Image” 요소
-        const noimg = document.createElement('div');
-        noimg.classList.add('noimage');
-        noimg.textContent = 'No Image';
-        cell.appendChild(noimg);
-
-        grid.appendChild(cell);
-      });
-    }
-
-    // 2) 이미지 로드 시도 함수: index.html과 이미지가 동일 폴더에 있을 때 파일명만 사용
-    function showCell(type, idx) {
-      const suffixIndex = getSuffixIndex(type, idx);
-      const prefix = prefixMap[type];
-      const imgId = `${prefix}_${suffixIndex}`;
-      const imgEl = document.getElementById(imgId);
-      if (!imgEl) {
-        console.warn(`showCell: element id="${imgId}" not found`);
-        return;
-      }
-      const cell = imgEl.parentElement;
-
-      // URL 구성: 동일 디렉토리에 있을 때 파일명만 사용. 
-      // 만약 서브폴더에 있다면 'waypoint_photos/${imgId}.png' 등으로 수정.
-      const url = `${imgId}.png?t=${Date.now()}`;
-      console.log('[showCell] trying load', imgId, url);
-
-      const tester = new Image();
-      tester.onload = () => {
-        imgEl.src = url;
-        cell.classList.add('ready');
-        cell.classList.remove('not-ready');
-        console.log('[showCell] loaded:', imgId);
-      };
-      tester.onerror = () => {
-        imgEl.src = '';
-        cell.classList.add('not-ready');
-        cell.classList.remove('ready');
-        console.log('[showCell] not found:', imgId);
-      };
-      tester.src = url;
-    }
-
-    // 3) 전체 갱신
-    function refreshGrid() {
-      for (let i = 1; i <= NUM_ENTRIES; i++) {
-        TYPES.forEach(type => {
-          showCell(type, i);
-        });
-      }
-    }
-
-    window.addEventListener('load', () => {
-      refreshGrid();
-      // 주기적 갱신: 5초마다
-      setInterval(refreshGrid, 5000);
-    });
-  </script>
-</body>
-</html>
