--- index.html
@@ -1,153 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8">
-  <title>inmo 교수님과 아이들</title>
-  <style>
-    @import url('https://fonts.googleapis.com/css2?family=Comic+Neue:wght@400;700&display=swap');
-    body {
-      font-family: 'Comic Neue', cursive;
-      background: linear-gradient(135deg, #FFDEE9 0%, #B5FFFC 100%);
-      margin: 0;
-      padding: 20px;
-      color: #333;
-    }
-    h1 {
-      text-align: center;
-      font-size: 2.5em;
-      margin-bottom: 30px;
-      color: #FF6F91;
-      text-shadow: 2px 2px 4px rgba(0,0,0,0.1);
-    }
-    #grid {
-      display: grid;
-      grid-template-columns: repeat(3, 1fr);
-      gap: 20px 10px;
-    }
-    .cell {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      background: #FFF;
-      border-radius: 12px;
-      box-shadow: 0 4px 6px rgba(0,0,0,0.1);
-      padding: 10px;
-      transition: transform 0.2s;
-    }
-    .cell:hover {
-      transform: translateY(-5px);
-    }
-    .cell h3 {
-      margin: 0 0 8px;
-      font-size: 1.1em;
-      opacity: 0.5;
-      transition: opacity .2s;
-      color: #555;
-    }
-    .cell img {
-      width: 100%;
-      max-height: 180px;
-      object-fit: contain;
-      border: 2px dashed #FF9FF3;
-      border-radius: 8px;
-      display: none;
-      background: #FFF;
-    }
-    .cell .noimage {
-      display: none;
-      color: #888;
-      font-size: 0.9em;
-    }
-    .cell.ready h3 { opacity: 1; color: #FF6F91; }
-    .cell.ready img { display: block; }
-    .cell.ready .noimage { display: none; }
-    .cell.not-ready h3 { opacity: 0.5; }
-    .cell.not-ready img { display: none; }
-    .cell.not-ready .noimage { display: block; }
-  </style>
-</head>
-<body>
-  <h1>inmo 교수님과 아이들</h1>
-  <div id="grid"></div>
-
-  <script>
-    const NUM_ENTRIES = 10;
-    const TYPES = ['Reference', 'Current', 'Diff'];
-    const grid = document.getElementById('grid');
-
-    // 1) 셀 생성
-    for (let i = 1; i <= NUM_ENTRIES; i++) {
-      TYPES.forEach(type => {
-        const cell = document.createElement('div');
-        cell.classList.add('cell', 'not-ready');
-
-        const title = document.createElement('h3');
-        title.textContent = `${type} ${i}`;
-        cell.appendChild(title);
-
-        const img = document.createElement('img');
-        img.id = `${type.toLowerCase()}${i}`;
-        img.alt = `${type} ${i}`;
-        cell.appendChild(img);
-
-        const noimg = document.createElement('div');
-        noimg.classList.add('noimage');
-        noimg.textContent = 'No Image';
-        cell.appendChild(noimg);
-
-        grid.appendChild(cell);
-      });
-    }
-
-    // 2) 파일 체크 & 상태 토글
-    async function showCell(type, idx) {
-      const key = type.toLowerCase();
-      const imgId = `${key}${idx}`;
-      const imgEl = document.getElementById(imgId);
-      const cell = imgEl.parentElement;
-
-      // 파일 이름 매핑
-      let filename;
-      if (type === 'Reference') {
-        filename = `reference${idx}.png`;
-      } else if (type === 'Current') {
-        filename = `waypoint_${idx}.png`;
-      } else {
-        filename = `comparsion_result_${idx}.png`;
-      }
-
-      const url = `waypoint_photos/${filename}?t=${Date.now()}`;
-
-      try {
-        const res = await fetch(url, { method: 'HEAD' });
-        if (res.ok) {
-          imgEl.src = url;
-          cell.classList.add('ready');
-          cell.classList.remove('not-ready');
-        } else {
-          cell.classList.add('not-ready');
-          cell.classList.remove('ready');
-        }
-      } catch {
-        cell.classList.add('not-ready');
-        cell.classList.remove('ready');
-      }
-    }
-
-    // 3) 전체 갱신
-    async function refreshGrid() {
-      for (let i = 1; i <= NUM_ENTRIES; i++) {
-        for (const type of TYPES) {
-          await showCell(type, i);
-        }
-      }
-    }
-
-    // 4) 초기 + 주기 갱신
-    window.addEventListener('load', () => {
-      refreshGrid();
-      setInterval(refreshGrid, 5000);
-    });
-  </script>
-</body>
-</html>
